Extract Workbench risk API module
