+++ docs/architecture/CODEBASE-REVIEW-LEDGER.md
@@ -116,3 +116,4 @@
 | LWB-R109 | Reporting API module extraction | Hardened | `src/features/workbench/api.ts` still owned report-batch materialization, status, run-once, archived-document metadata, and document-download URL helpers after the reporting operations hook had already been extracted. That kept report/archive request shaping coupled to unrelated Workbench API families and made Gateway reporting changes noisier to review. | Added `reporting-api.ts` for report-batch and archived-document Gateway helpers, reusing `reporting-api-headers.ts`, with `api.ts` re-exporting the same public functions for compatibility. Preserved explicit-portfolio report-batch payloads, source-owned status/run-once posture, archive metadata retrieval, idempotency, and observability labels without adding browser-side batch membership, report generation truth, archive truth, retention decisions, delivery, communication, OMS, or execution behavior. | Continue splitting `api.ts` by coherent route family only where it removes real review coupling. Preserve the public `api.ts` barrel until downstream imports are migrated deliberately. | `src/features/workbench/reporting-api.ts`, `src/features/workbench/reporting-api-headers.ts`, `src/features/workbench/api.ts`; validation via Workbench API reporting tests, report-batch operations hook tests, `npm run typecheck`, `npm run lint`, `git diff --check`, `npm run build`. |
 | LWB-R110 | Construction API module extraction | Hardened | `src/features/workbench/api.ts` still owned construction alternative generation, alternative-set retrieval, source-owned selection, external OMS acknowledgement supportability, construction idempotency, and construction request shaping after the construction action hook had already been extracted. That kept construction Gateway route changes coupled to unrelated Workbench API families. | Added `construction-api.ts` for construction Gateway helpers, source-product acknowledgement retrieval, construction caller headers, request shaping, and bounded idempotency/correlation generation, with `api.ts` re-exporting the same public functions for compatibility. Preserved Gateway-shaped stateful construction selectors and source-owned acknowledgement supportability without adding optimizer logic, price construction, selection truth, OMS acknowledgement truth, fills, settlement, execution status, or autonomous execution behavior. | Continue splitting `api.ts` by coherent route family only where it removes real review coupling. Preserve the public `api.ts` barrel until downstream imports are migrated deliberately. | `src/features/workbench/construction-api.ts`, `src/features/workbench/api.ts`; validation via Workbench API construction tests, construction action hook tests, `npm run typecheck`, `npm run lint`, `git diff --check`, `npm run build`. |
 | LWB-R111 | DPM wave API module extraction | Hardened | `src/features/workbench/api.ts` still owned rebalance-wave and campaign-definition Gateway helpers after DPM wave action orchestration had already moved to `use-dpm-wave-command-center-actions.ts`. That kept campaign workflow request shaping, launch-readiness calls, supportability reads, report-input reads, and AI handoff helpers coupled to unrelated Workbench API families. | Added `dpm-wave-api.ts` for DPM wave and campaign-definition list/detail/actions/readiness/launch/supportability/report-input/AI handoff helpers, with `api.ts` re-exporting the same public functions for compatibility. Preserved Gateway-shaped payloads, existing caller headers, observability operation names, and source-owned campaign workflow posture. | Continue splitting `api.ts` by coherent Gateway route family only where it removes real review coupling. Keep campaign cohort discovery, membership, readiness calculation, idempotency reconstruction beyond the existing helper, trade approval, order generation, routing, OMS, fills, settlement, execution, and maker-checker claims out of Workbench. | `src/features/workbench/dpm-wave-api.ts`, `src/features/workbench/api.ts`; validation via Workbench API DPM wave tests, DPM wave hook tests, analytics observability tests, `npm run typecheck`, `npm run lint`, `git diff --check`, `npm run build`. |
+| LWB-R112 | Risk API module extraction | Hardened | `src/features/workbench/api.ts` still owned risk summary, concentration, drawdown, rolling, and attribution Gateway helpers plus risk query shaping after multiple DPM and reporting API families had already been split out. That kept risk route changes coupled to unrelated portfolio, performance, sandbox, reporting, and DPM helpers. | Added `risk-api.ts` for Gateway-backed Workbench risk helpers and risk query construction, with `api.ts` re-exporting the same public functions for compatibility. Preserved existing Gateway paths, query parameters, correlation headers, and bounded observability operation names. | Continue extracting coherent route families from `api.ts` only where ownership becomes clearer. Keep volatility, drawdown, concentration, rolling-risk, attribution, freshness, and supportability truth source-owned; do not calculate or reinterpret risk analytics in browser components. | `src/features/workbench/risk-api.ts`, `src/features/workbench/api.ts`; validation via Workbench API risk tests, Performance/Risk mode tests, analytics observability tests, `npm run typecheck`, `npm run lint`, `git diff --check`, `npm run build`. |
